+++ tesseract/tesseract_planning/tesseract_motion_planners/include/tesseract_motion_planners/descartes/impl/descartes_external_positioner_sampler.hpp
@@ -123,7 +123,7 @@ void DescartesExternalPositionerSampler<FloatType>::nested_ik(
     bool get_best_solution,
     double& distance)
 {
-  if (loop_level >= positioner_kinematics_->numJoints())
+  if (loop_level >= static_cast<int>(positioner_kinematics_->numJoints()))
   {
     ikAt(solution_set, target_pose, sample_pose, get_best_solution, distance);
     return;
@@ -154,7 +154,7 @@ bool DescartesExternalPositionerSampler<FloatType>::ikAt(
     return false;
 
   Eigen::VectorXd robot_solution_set;
-  int robot_dof = robot_kinematics_->numJoints();
+  int robot_dof = static_cast<int>(robot_kinematics_->numJoints());
   if (!robot_kinematics_->calcInvKin(robot_solution_set, robot_target_pose, ik_seed_))
     return false;
 
@@ -168,7 +168,7 @@ bool DescartesExternalPositionerSampler<FloatType>::ikAt(
     full_sol.insert(end(full_sol), std::make_move_iterator(sol), std::make_move_iterator(sol + robot_dof));
 
     if ((is_valid_ != nullptr) &&
-        !is_valid_(Eigen::Map<Eigen::Matrix<FloatType, Eigen::Dynamic, 1>>(full_sol.data(), full_sol.size())))
+        !is_valid_(Eigen::Map<Eigen::Matrix<FloatType, Eigen::Dynamic, 1>>(full_sol.data(), static_cast<long>(full_sol.size()))))
       continue;
 
     if (!get_best_solution)
